requirements.txt
mysql-connector-python==8.1.0
pytest==7.4.2
requests==2.31.0
selenium==4.13.0